backend/requirements.txt
asgiref==3.9.1
django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
sqlparse==0.5.3
